Fix: renamed PopulationVar to PopulationVariance
