+++ Statistics/Statistics.py
@@ -1,7 +1,7 @@
 from Calculators.Calculator import Calculator
 from Statistics.ZScore import z_score
 from Statistics.ConfidenceInterval import confidence_interval
-from Statistics.PopulationVar import population_variance
+from Statistics.PopulationVariance import population_variance
 from Statistics.PopulationMean import population_mean
 from Statistics.Median import median
 from Statistics.Mode import mode
